branch load into foreground and background work
