+++ x-pack/platform/plugins/shared/alerting/server/application/rule/methods/bulk_create/bulk_create_rules.ts
@@ -5,121 +5,44 @@
  * 2.0.
  */
 
-import Boom from '@hapi/boom';
-import Semver from 'semver';
 import pMap from 'p-map';
-import { i18n } from '@kbn/i18n';
 import { withSpan } from '@kbn/apm-utils';
-import type {
-  SavedObject,
-  SavedObjectReference,
-  SavedObjectsBulkCreateObject,
-} from '@kbn/core/server';
+import type { SavedObject, SavedObjectsBulkCreateObject } from '@kbn/core/server';
 import { SavedObjectsUtils } from '@kbn/core/server';
-import type { TaskInstanceWithDeprecatedFields } from '@kbn/task-manager-plugin/server/task';
 
-import { validateAndAuthorizeSystemActions } from '../../../../lib/validate_authorize_system_actions';
 import { RULE_SAVED_OBJECT_TYPE } from '../../../../saved_objects';
-import { parseDuration, getRuleCircuitBreakerErrorMessage } from '../../../../../common';
-import { WriteOperations, AlertingAuthorizationEntity } from '../../../../authorization';
-import {
-  validateRuleTypeParams,
-  getRuleNotifyWhenType,
-  getDefaultMonitoringRuleDomainProperties,
-} from '../../../../lib';
-import { getRuleExecutionStatusPending } from '../../../../lib/rule_execution_status';
-import {
-  addGeneratedActionValues,
-  createNewAPIKeySet,
-  extractReferences,
-  updateMeta,
-  validateActions,
-} from '../../../../rules_client/lib';
-import {
-  apiKeyAsAlertAttributes,
-  apiKeyAsRuleDomainProperties,
-} from '../../../../rules_client/common';
+import { getRuleCircuitBreakerErrorMessage } from '../../../../../common';
+import { updateMeta } from '../../../../rules_client/lib';
 import { API_KEY_GENERATE_CONCURRENCY } from '../../../../rules_client/common/constants';
 import { ruleAuditEvent, RuleAuditAction } from '../../../../rules_client/common/audit_events';
-import { tryToEnableTasks } from '../bulk_enable/bulk_enable_rules';
-import { bulkMarkApiKeysForInvalidation } from '../../../../invalidate_pending_api_keys/bulk_mark_api_keys_for_invalidation';
 import { bulkCreateRulesSo } from '../../../../data/rule';
-import type { RawRule, RuleTypeRegistry, SanitizedRule } from '../../../../types';
-import type { IntervalSchedule } from '../../../../../common';
+import type { RawRule, SanitizedRule } from '../../../../types';
 import type { BulkOperationError, RulesClientContext } from '../../../../rules_client/types';
-import type { RuleDomain, RuleParams } from '../../types';
-import {
-  transformRuleAttributesToRuleDomain,
-  transformRuleDomainToRule,
-  transformRuleDomainToRuleAttributes,
-} from '../../transforms';
-import { ruleDomainSchema } from '../../schemas';
-import { createRuleDataSchema } from '../create/schemas';
+import type { RuleParams } from '../../types';
 import { validateScheduleLimit } from '../get_schedule_frequency';
 import type {
   BulkCreateOperationError,
-  BulkCreateDisabledReason,
-  BulkCreateRulesItem,
   BulkCreateRulesParams,
   BulkCreateRulesResult,
 } from './types';
+import {
+  buildTaskInstance,
+  collectNewKeysToInvalidate,
+  demotePersistedRules,
+  flushKeysToInvalidate,
+  prepareRule,
+  toSanitizedRule,
+} from './utils';
+import type { ApiKeyEntry, DemotionEntry, PreparedRule } from './utils';
 
-const getBulkCreateAsDisabledMessage = (message: string): string =>
-  i18n.translate('xpack.alerting.rulesClient.bulkCreate.ruleCreatedDisabledErrorMessage', {
-    defaultMessage: 'Rule created in a disabled state: {message}',
-    values: { message },
-  });
-
-interface PreparedRule {
-  id: string;
-  name: string;
-  enabled: boolean;
-  rawRule: RawRule;
-  references: SavedObjectReference[];
-  schedule: IntervalSchedule;
-  consumer: string;
-  ruleTypeId: string;
-}
-
-interface ApiKeyEntry {
-  apiKey: string | null;
-  uiamApiKey: string | null;
-  apiKeyCreatedByUser: boolean | null;
-}
-
-const collectNewKeysToInvalidate = (entries: Iterable<ApiKeyEntry>): string[] => {
-  const keys: string[] = [];
-  for (const { apiKey, uiamApiKey, apiKeyCreatedByUser } of entries) {
-    if (apiKey && !apiKeyCreatedByUser) keys.push(apiKey);
-    if (uiamApiKey && !apiKeyCreatedByUser) keys.push(uiamApiKey);
-  }
-  return keys;
-};
-
-const buildTaskInstance = (
-  context: RulesClientContext,
-  prepared: PreparedRule
-): TaskInstanceWithDeprecatedFields => ({
-  id: prepared.id,
-  taskType: `alerting:${prepared.ruleTypeId}`,
-  schedule: prepared.schedule,
-  params: {
-    alertId: prepared.id,
-    spaceId: context.spaceId,
-    consumer: prepared.consumer,
-  },
-  state: {
-    previousStartedAt: null,
-    alertTypeState: {},
-    alertInstances: {},
-  },
-  scope: ['alerting'],
-  // Tasks are scheduled disabled. Phase 5 enables them via taskManager.bulkEnable
-  // so per-task validation drops never produce a running task without a rule SO,
-  // and the activation can randomise schedule datetimes across the batch.
-  enabled: false,
-});
-
+/**
+ * Persist-first bulk rule create. 2 parts:
+ * 1. Foreground: SO bulkCreate, audit, return.
+ * 2. Background (`result.backgroundWork`): Scheduling - limit checks, task scheduling, demotion / rule SO update (-> disabled), key invalidation.
+ *
+ * Returned `rules[]` reflect input intent; the background promise may
+ * later demote rules to "disabled" if related tasks failed to schedule.
+ */
 export async function bulkCreateRules<Params extends RuleParams = never>(
   context: RulesClientContext,
   params: BulkCreateRulesParams<Params>
@@ -129,7 +52,7 @@ export async function bulkCreateRules<Params extends RuleParams = never>(
   const total = rules.length;
 
   if (total === 0) {
-    return { rules: [], errors: [], total: 0, taskIdsFailedToBeEnabled: [] };
+    return { rules: [], errors: [], total: 0, backgroundWork: Promise.resolve([]) };
   }
 
   const username = await context.getUserName();
@@ -147,7 +70,6 @@ export async function bulkCreateRules<Params extends RuleParams = never>(
   const preparedRules = new Map<string, PreparedRule>();
   const errors: BulkOperationError[] = [];
   const apiKeysMap = new Map<string, ApiKeyEntry>();
-  // Key invalidation involves call to ES so we invalidate keys at the end to reduce round-trips.
   const keysToInvalidate = new Set<string>();
 
   await pMap(
@@ -169,100 +91,9 @@ export async function bulkCreateRules<Params extends RuleParams = never>(
     { concurrency: API_KEY_GENERATE_CONCURRENCY }
   );
 
-  // Phase 2: validate schedule-limits, enabled subset only.
-  const enabled = [...preparedRules.values()].filter((p) => p.enabled);
-
-  if (enabled.length > 0) {
-    const updatedInterval = enabled.map((r) => r.schedule.interval);
-    const validationPayload = await validateScheduleLimit({ context, updatedInterval });
-    const enabledIds = enabled.map((p) => p.id);
-    if (validationPayload) {
-      const reasonMessage = getRuleCircuitBreakerErrorMessage({
-        interval: validationPayload.interval,
-        intervalAvailable: validationPayload.intervalAvailable,
-        action: 'bulkCreate',
-        rules: enabledIds.length,
-      });
-      logger.warn(`Demoting ${enabledIds.length} rules -> disabled, schedule limit exceeded.`);
-      demotePreparedRules({
-        ids: enabledIds,
-        reason: 'schedule_limit_exceeded',
-        message: reasonMessage,
-        preparedRules,
-        apiKeysMap,
-        keysToInvalidate,
-        errors,
-        username,
-      });
-    }
-  }
-
-  // Phase 3: schedule tasks for the surviving enabled subset.
-  const survivingEnabled = [...preparedRules.values()].filter((p) => p.enabled);
-  const newlyScheduledTaskIds = new Set<string>();
-
-  if (survivingEnabled.length > 0) {
-    const tasksToSchedule = survivingEnabled.map((preparedRule) =>
-      buildTaskInstance(context, preparedRule)
-    );
-
-    let scheduledIds: string[] = [];
-    const survivingEnabledIds = survivingEnabled.map((p) => p.id);
-    try {
-      const scheduledTasks = await withSpan(
-        { name: 'taskManager.bulkSchedule', type: 'tasks' },
-        () => context.taskManager.bulkSchedule(tasksToSchedule)
-      );
-      scheduledIds = scheduledTasks.map((task) => task.id);
-    } catch (error) {
-      // Whole-call TM throw: demote enabled subset to disabled, continue.
-      logger.warn(
-        `Demoting ${survivingEnabledIds.length} rules -> disabled, task scheduling failed.`
-      );
-      demotePreparedRules({
-        ids: survivingEnabledIds,
-        reason: 'task_schedule_failed',
-        message: `Failed to schedule tasks: ${error.message}`,
-        preparedRules,
-        apiKeysMap,
-        keysToInvalidate,
-        errors,
-        username,
-      });
-    }
-
-    if (scheduledIds.length > 0) {
-      scheduledIds.forEach((id) => newlyScheduledTaskIds.add(id));
-    }
-
-    // Silent per-task drops: bulkSchedule's `taskInstanceToAttributes` validation
-    // logs+skips invalid instances. Diff requested vs returned and demote the
-    // missing ones — re-enabling later will hit the same validation, but at
-    // least the rule definition isn't lost.
-    if (preparedRules.size > 0 && scheduledIds.length < survivingEnabledIds.length) {
-      const returned = new Set(scheduledIds);
-      const dropped = survivingEnabledIds.filter((id) => !returned.has(id));
-      if (dropped.length > 0) {
-        logger.warn(`Demoting ${dropped.length} rules -> disabled, task validation failed.`);
-        demotePreparedRules({
-          ids: dropped,
-          reason: 'task_validation_failed',
-          message: 'Task scheduling silently dropped this rule (validation failure in task store)',
-          preparedRules,
-          apiKeysMap,
-          keysToInvalidate,
-          errors,
-          username,
-        });
-      }
-    }
-  }
-
-  // No survivors at all: still flush any pending key invalidations from
-  // Phase 1 demotions (no keys to flush in that case, but keep it explicit).
   if (preparedRules.size === 0) {
     await flushKeysToInvalidate(keysToInvalidate, context);
-    return { rules: [], errors, total, taskIdsFailedToBeEnabled: [] };
+    return { rules: [], errors, total, backgroundWork: Promise.resolve([]) };
   }
 
   // Audit per-rule CREATE event before persistence (mirrors createRuleSavedObject).
@@ -276,7 +107,7 @@ export async function bulkCreateRules<Params extends RuleParams = never>(
     );
   }
 
-  // Phase 4: bulk SO create (no overwrite — id collisions surface as per-row 409)
+  // Phase 2: bulk SO create (no overwrite — id collisions surface as per-row 409).
   const bulkObjects: Array<SavedObjectsBulkCreateObject<RawRule>> = [...preparedRules.values()].map(
     (prepared) => ({
       type: RULE_SAVED_OBJECT_TYPE,
@@ -297,32 +128,14 @@ export async function bulkCreateRules<Params extends RuleParams = never>(
         })
     );
   } catch (error) {
-    // Whole-call SO failure (auth, timeout, etc): invalidate every newly-minted
-    // key (those tracked in `apiKeysMap`, plus anything previous phases queued
-    // into `keysToInvalidate`), best-effort orphan-task cleanup scoped to ids we
-    // actually scheduled in Phase 3, then rethrow. Inline flush — control flow
-    // never reaches the end-of-function flush.
+    // Nothing was scheduled — invalidate minted keys and rethrow.
     for (const k of collectNewKeysToInvalidate(apiKeysMap.values())) keysToInvalidate.add(k);
     await flushKeysToInvalidate(keysToInvalidate, context);
-    if (newlyScheduledTaskIds.size > 0) {
-      try {
-        await context.taskManager.bulkRemove([...newlyScheduledTaskIds]);
-      } catch (cleanupError) {
-        logger.error(
-          `bulkCreateRules: failed to clean up tasks ${[...newlyScheduledTaskIds].join(
-            ', '
-          )} after SO bulkCreate threw: ${cleanupError.message}`
-        );
-      }
-    }
     throw error;
   }
 
-  // Phase 4 per-row outcomes.
+  // Phase 3: per-row outcome split.
   const successfulSos: Array<SavedObject<RawRule>> = [];
-  const taskIdsToEnable: string[] = [];
-  const taskIdsToCleanUp: string[] = [];
-
   for (const so of bulkResponse.saved_objects) {
     if (so.error) {
       errors.push({
@@ -334,18 +147,12 @@ export async function bulkCreateRules<Params extends RuleParams = never>(
       const apiKey = apiKeysMap.get(so.id);
       if (apiKey) {
         for (const k of collectNewKeysToInvalidate([apiKey])) keysToInvalidate.add(k);
-      }
-
-      // Only ids we scheduled in Phase 3. Skipping caller-supplied id collisions
-      // avoids nuking a pre-existing rule's task on a 409.
-      if (newlyScheduledTaskIds.has(so.id)) {
-        taskIdsToCleanUp.push(so.id);
+        apiKeysMap.delete(so.id);
       }
     } else {
       successfulSos.push(so as SavedObject<RawRule>);
-      if (newlyScheduledTaskIds.has(so.id)) {
-        taskIdsToEnable.push(so.id);
-        // Audit per-rule ENABLE for the enabled subset (mirrors single-rule semantics).
+      if (so.attributes.enabled) {
+        // Background may later demote this rule and emit a DISABLE audit.
         context.auditLogger?.log(
           ruleAuditEvent({
             action: RuleAuditAction.ENABLE,
@@ -361,334 +168,112 @@ export async function bulkCreateRules<Params extends RuleParams = never>(
     }
   }
 
-  // Single batched TM cleanup for per-row failures.
-  if (taskIdsToCleanUp.length > 0) {
+  const enabledPersistedIds = successfulSos
+    .filter((so) => so.attributes.enabled)
+    .map((so) => so.id);
+
+  // Phase 4: detached background work — wrapped so the promise never rejects.
+  // Resolves to one error per demoted rule (empty array on full success).
+  const backgroundWork = (async (): Promise<BulkCreateOperationError[]> => {
+    const bgErrors: BulkCreateOperationError[] = [];
     try {
-      logger.warn(`Cleaning up ${taskIdsToCleanUp.length} tasks where SO creation failed.`);
-      await context.taskManager.bulkRemove(taskIdsToCleanUp);
-    } catch (cleanupError) {
-      logger.error(
-        `bulkCreateRules: failed to clean up tasks ${taskIdsToCleanUp.join(
-          ', '
-        )} after SO per-row errors: ${cleanupError.message}`
-      );
+      const demotedIds = new Map<string, DemotionEntry>();
+
+      // Phase 4A: schedule-limit check.
+      if (enabledPersistedIds.length > 0) {
+        const updatedInterval = enabledPersistedIds.map(
+          (id) => preparedRules.get(id)!.schedule.interval
+        );
+        const validationPayload = await validateScheduleLimit({ context, updatedInterval });
+        if (validationPayload) {
+          const reasonMessage = getRuleCircuitBreakerErrorMessage({
+            interval: validationPayload.interval,
+            intervalAvailable: validationPayload.intervalAvailable,
+            action: 'bulkCreate',
+            rules: enabledPersistedIds.length,
+          });
+          logger.warn(
+            `Demoting ${enabledPersistedIds.length} rules -> disabled, schedule limit exceeded.`
+          );
+          for (const id of enabledPersistedIds) {
+            demotedIds.set(id, { reason: 'schedule_limit_exceeded', message: reasonMessage });
+          }
+        }
+      }
+
+      // Phase 4B: schedule tasks for ids not already demoted by Phase 4A.
+      const idsToSchedule = enabledPersistedIds.filter((id) => !demotedIds.has(id));
+      if (idsToSchedule.length > 0) {
+        const tasksToSchedule = idsToSchedule.map((id) =>
+          buildTaskInstance(context, preparedRules.get(id)!)
+        );
+
+        let scheduledIds: string[] = [];
+        let bulkScheduleThrew = false;
+        try {
+          const scheduledTasks = await withSpan(
+            { name: 'taskManager.bulkSchedule', type: 'tasks' },
+            () => context.taskManager.bulkSchedule(tasksToSchedule)
+          );
+          scheduledIds = scheduledTasks.map((task) => task.id);
+        } catch (error) {
+          // Whole-call TM throw: demote everything we tried to schedule.
+          bulkScheduleThrew = true;
+          logger.warn(
+            `Demoting ${idsToSchedule.length} rules -> disabled, task scheduling failed.`
+          );
+          for (const id of idsToSchedule) {
+            demotedIds.set(id, {
+              reason: 'task_schedule_failed',
+              message: `Failed to schedule tasks: ${error.message}`,
+            });
+          }
+        }
+
+        // Silent per-task drops: bulkSchedule logs+skips invalid instances.
+        // Skip when the whole call threw — those ids are already accounted for
+        // under `task_schedule_failed` and must not be reclassified.
+        if (!bulkScheduleThrew && scheduledIds.length < idsToSchedule.length) {
+          const returned = new Set(scheduledIds);
+          const dropped = idsToSchedule.filter((id) => !returned.has(id));
+          if (dropped.length > 0) {
+            logger.warn(`Demoting ${dropped.length} rules -> disabled, task validation failed.`);
+            for (const id of dropped) {
+              demotedIds.set(id, {
+                reason: 'task_schedule_entry_failed',
+                message:
+                  'Task scheduling silently dropped this rule (validation failure in task store)',
+              });
+            }
+          }
+        }
+      }
+
+      // Phase 4C: persist demotions (bulkUpdate SOs to disabled).
+      if (demotedIds.size > 0) {
+        const demotionErrors = await demotePersistedRules({
+          context,
+          demotedIds,
+          preparedRules,
+          apiKeysMap,
+          keysToInvalidate,
+          username,
+        });
+        bgErrors.push(...demotionErrors);
+      }
+
+      // Phase 4D: flush queued API key invalidations.
+      await flushKeysToInvalidate(keysToInvalidate, context);
+    } catch (err) {
+      logger.error(`bulkCreateRules background phases failed: ${err.message}`);
     }
-  }
+    return bgErrors;
+  })();
 
-  // Phase 5: enable tasks for successfully persisted enabled rules.
-  const taskIdsFailedToBeEnabled = await tryToEnableTasks({
-    taskIdsToEnable,
-    logger,
-    taskManager: context.taskManager,
-  });
-
-  // Single end-of-function flush for all collected key invalidations.
-  await flushKeysToInvalidate(keysToInvalidate, context);
-
-  // Phase 6: domain transform + return.
+  // Phase 5: domain transform + return (reflects input intent).
   const sanitizedRules: Array<SanitizedRule<Params>> = successfulSos.map((so) =>
     toSanitizedRule<Params>(context, so, context.ruleTypeRegistry)
   );
 
-  return {
-    rules: sanitizedRules,
-    errors,
-    total,
-    taskIdsFailedToBeEnabled, // <-- same as bulkEnableRules().
-  };
+  return { rules: sanitizedRules, errors, total, backgroundWork };
 }
-
-const toSanitizedRule = <Params extends RuleParams = never>(
-  context: RulesClientContext,
-  so: SavedObject<RawRule>,
-  ruleTypeRegistry: RuleTypeRegistry
-): SanitizedRule<Params> => {
-  const ruleType = ruleTypeRegistry.get(so.attributes.alertTypeId);
-  const ruleDomain: RuleDomain<Params> = transformRuleAttributesToRuleDomain<Params>(
-    so.attributes,
-    {
-      id: so.id,
-      logger: context.logger,
-      ruleType,
-      references: so.references,
-      omitGeneratedValues: false,
-    },
-    context.isSystemAction
-  );
-
-  try {
-    ruleDomainSchema.validate(ruleDomain);
-  } catch (e) {
-    context.logger.warn(`Error validating bulk-created rule domain object for id: ${so.id}, ${e}`);
-  }
-
-  return transformRuleDomainToRule<Params>(ruleDomain, { isPublic: true }) as SanitizedRule<Params>;
-};
-
-interface PrepareRuleArgs<Params extends RuleParams> {
-  context: RulesClientContext;
-  actionsClient: Awaited<ReturnType<RulesClientContext['getActionsClient']>>;
-  username: string | null;
-  id: string;
-  rule: BulkCreateRulesItem<Params>;
-  authzCache: Map<string, Promise<void>>;
-  errors: BulkCreateOperationError[];
-  apiKeysMap: Map<string, ApiKeyEntry>;
-}
-
-const prepareRule = async <Params extends RuleParams>({
-  context,
-  actionsClient,
-  username,
-  id,
-  rule,
-  authzCache,
-  errors,
-  apiKeysMap,
-}: PrepareRuleArgs<Params>): Promise<{ prepared?: PreparedRule; error?: BulkOperationError }> => {
-  const { allowMissingConnectorSecrets } = rule;
-
-  try {
-    const { actions: genActions, systemActions: genSystemActions } = await addGeneratedActionValues(
-      rule.data.actions,
-      rule.data.systemActions,
-      context
-    );
-    const data = { ...rule.data, actions: genActions, systemActions: genSystemActions };
-
-    try {
-      createRuleDataSchema.validate(data);
-    } catch (validationError) {
-      throw Boom.badRequest(`Error validating create data - ${validationError.message}`);
-    }
-
-    // ruleTypeRegistry.get throws 400 if not registered.
-    context.ruleTypeRegistry.get(data.alertTypeId);
-
-    const authzKey = `${data.alertTypeId}::${data.consumer}`;
-    if (!authzCache.has(authzKey)) {
-      authzCache.set(
-        authzKey,
-        context.authorization.ensureAuthorized({
-          ruleTypeId: data.alertTypeId,
-          consumer: data.consumer,
-          operation: WriteOperations.Create,
-          entity: AlertingAuthorizationEntity.Rule,
-        })
-      );
-    }
-    try {
-      await authzCache.get(authzKey)!;
-    } catch (authzError) {
-      context.auditLogger?.log(
-        ruleAuditEvent({
-          action: RuleAuditAction.CREATE,
-          savedObject: { type: RULE_SAVED_OBJECT_TYPE, id, name: data.name },
-          error: authzError,
-        })
-      );
-      throw authzError;
-    }
-
-    context.ruleTypeRegistry.ensureRuleTypeEnabled(data.alertTypeId);
-    const ruleType = context.ruleTypeRegistry.get(data.alertTypeId);
-    const validatedRuleTypeParams = validateRuleTypeParams(data.params, ruleType.validate.params);
-
-    await validateActions(context, ruleType, data, allowMissingConnectorSecrets);
-    await validateAndAuthorizeSystemActions({
-      actionsClient,
-      actionsAuthorization: context.actionsAuthorization,
-      connectorAdapterRegistry: context.connectorAdapterRegistry,
-      systemActions: data.systemActions ?? [],
-      rule: { consumer: data.consumer, producer: ruleType.producer },
-    });
-
-    const intervalInMs = parseDuration(data.schedule.interval);
-    if (
-      intervalInMs < context.minimumScheduleIntervalInMs &&
-      context.minimumScheduleInterval.enforce
-    ) {
-      throw Boom.badRequest(
-        `Error creating rule: the interval is less than the allowed minimum interval of ${context.minimumScheduleInterval.value}`
-      );
-    }
-    if (
-      intervalInMs < context.minimumScheduleIntervalInMs &&
-      !context.minimumScheduleInterval.enforce
-    ) {
-      context.logger.warn(
-        `Rule schedule interval (${data.schedule.interval}) for "${ruleType.id}" rule type with ID "${id}" is less than the minimum value (${context.minimumScheduleInterval.value}). Running rules at this interval may impact alerting performance. Set "xpack.alerting.rules.minimumScheduleInterval.enforce" to true to prevent creation of these rules.`
-      );
-    }
-
-    // Mint API key for enabled rules.
-    // Soft-fail: a key-mint failure does NOT reject the rule. We persist it as
-    // disabled, push a degraded error so the caller knows, and let downstream
-    // phases treat it as a never-enabled rule. Re-enabling later will re-mint.
-    let effectiveEnabled = data.enabled;
-    let apiKeyProps:
-      | ReturnType<typeof apiKeyAsRuleDomainProperties>
-      | Awaited<ReturnType<typeof createNewAPIKeySet>> = apiKeyAsRuleDomainProperties(
-      null,
-      username,
-      false
-    );
-    if (data.enabled) {
-      try {
-        apiKeyProps = await createNewAPIKeySet(context, {
-          id: ruleType.id,
-          ruleName: data.name,
-          username,
-          shouldUpdateApiKey: true,
-          errorMessage: 'Error creating rule: could not create API key',
-        });
-        apiKeysMap.set(id, {
-          apiKey: apiKeyProps.apiKey ?? null,
-          uiamApiKey: apiKeyProps.uiamApiKey ?? null,
-          apiKeyCreatedByUser: apiKeyProps.apiKeyCreatedByUser ?? null,
-        });
-      } catch (apiKeyErr) {
-        effectiveEnabled = false;
-        apiKeyProps = apiKeyAsRuleDomainProperties(null, username, false);
-        errors.push({
-          message: getBulkCreateAsDisabledMessage(apiKeyErr.message),
-          status: apiKeyErr.output?.statusCode,
-          rule: { id, name: data.name },
-          disabledReason: 'api_key_creation_failed',
-        });
-      }
-    }
-
-    const allActions = [...data.actions, ...(data.systemActions ?? [])];
-    const artifacts = data.artifacts ?? {};
-    const {
-      references,
-      params: updatedParams,
-      actions: actionsWithRefs,
-      artifacts: artifactsWithRefs,
-    } = await extractReferences(context, ruleType, allActions, validatedRuleTypeParams, artifacts);
-
-    const createTime = Date.now();
-    const lastRunTimestamp = new Date();
-    const legacyId = Semver.lt(context.kibanaVersion, '8.0.0') ? id : null;
-    const notifyWhen = getRuleNotifyWhenType(data.notifyWhen ?? null, data.throttle ?? null);
-    const throttle = data.throttle ?? null;
-    const { systemActions: _sa, actions: _a, ...restData } = data;
-
-    const ruleAttributes = transformRuleDomainToRuleAttributes({
-      actionsWithRefs,
-      artifactsWithRefs,
-      rule: {
-        ...restData,
-        ...apiKeyProps,
-        enabled: effectiveEnabled,
-        id,
-        createdBy: username,
-        updatedBy: username,
-        createdAt: new Date(createTime),
-        updatedAt: new Date(createTime),
-        snoozeSchedule: [],
-        muteAll: false,
-        mutedInstanceIds: [],
-        notifyWhen,
-        throttle,
-        executionStatus: getRuleExecutionStatusPending(lastRunTimestamp.toISOString()),
-        monitoring: getDefaultMonitoringRuleDomainProperties(lastRunTimestamp.toISOString()),
-        revision: 0,
-        running: false,
-      } as unknown as RuleDomain<Params>,
-      params: { legacyId, paramsWithRefs: updatedParams },
-    });
-
-    if (effectiveEnabled) {
-      ruleAttributes.lastEnabledAt = new Date(createTime).toISOString();
-      ruleAttributes.scheduledTaskId = id;
-    }
-
-    const prepared = {
-      id,
-      name: data.name,
-      enabled: effectiveEnabled,
-      rawRule: ruleAttributes,
-      references,
-      schedule: data.schedule,
-      consumer: data.consumer,
-      ruleTypeId: data.alertTypeId,
-    };
-    return { prepared };
-  } catch (err) {
-    const error = {
-      message: err.message,
-      status: err.output?.statusCode,
-      rule: { id, name: rule.data?.name ?? 'n/a' },
-    };
-    return { error };
-  }
-};
-
-/**
- * Demote in-memory (enabled -> disabled): flips a set of currently-enabled
- * prepared rules to disabled, queues their API keys for invalidation, records a degraded
- * error so the caller can surface "rule was created in a disabled state".
- */
-const demotePreparedRules = ({
-  ids,
-  reason,
-  message,
-  preparedRules,
-  apiKeysMap,
-  keysToInvalidate,
-  errors,
-  username,
-}: {
-  ids: string[];
-  reason: BulkCreateDisabledReason;
-  message: string;
-  preparedRules: Map<string, PreparedRule>;
-  apiKeysMap: Map<string, ApiKeyEntry>;
-  keysToInvalidate: Set<string>;
-  errors: BulkCreateOperationError[];
-  username: string | null;
-}): void => {
-  for (const id of ids) {
-    const prepared = preparedRules.get(id);
-    if (!prepared || !prepared.enabled) continue;
-
-    const apiKey = apiKeysMap.get(id);
-    if (apiKey) {
-      for (const k of collectNewKeysToInvalidate([apiKey])) keysToInvalidate.add(k);
-      apiKeysMap.delete(id);
-    }
-
-    // Re-shape `rawRule` to the disabled-rule form.
-    const nullKey = apiKeyAsAlertAttributes(null, username, false);
-    prepared.rawRule = {
-      ...prepared.rawRule,
-      ...nullKey,
-      uiamApiKey: null,
-      enabled: false,
-    };
-    delete prepared.rawRule.scheduledTaskId;
-    delete prepared.rawRule.lastEnabledAt;
-    prepared.enabled = false;
-
-    errors.push({
-      message: getBulkCreateAsDisabledMessage(message),
-      rule: { id, name: prepared.name },
-      disabledReason: reason,
-    });
-  }
-};
-
-const flushKeysToInvalidate = async (
-  keysToInvalidate: Set<string>,
-  context: RulesClientContext
-): Promise<void> => {
-  if (keysToInvalidate.size === 0) return;
-  await bulkMarkApiKeysForInvalidation(
-    { apiKeys: [...keysToInvalidate] },
-    context.logger,
-    context.unsecuredSavedObjectsClient
-  );
-  keysToInvalidate.clear();
-};
