+++ x-pack/solutions/security/plugins/security_solution/server/lib/detection_engine/rule_management/logic/detection_rules_client/detection_rules_client.bulk_create_prebuilt_rules.test.ts
@@ -55,7 +55,7 @@ describe('DetectionRulesClient.bulkCreatePrebuiltRules', () => {
       rules: [getRuleMock(getQueryRuleParams()), getRuleMock(getQueryRuleParams())],
       errors: [],
       total: 2,
-      taskIdsFailedToBeEnabled: [],
+      backgroundWork: Promise.resolve([]),
     });
 
     const { results, errors } = await detectionRulesClient.bulkCreatePrebuiltRules({
@@ -80,7 +80,7 @@ describe('DetectionRulesClient.bulkCreatePrebuiltRules', () => {
         rules: [],
         errors: [{ message: 'boom', status: 500, rule: { id, name: asset.name } }],
         total: 1,
-        taskIdsFailedToBeEnabled: [],
+        backgroundWork: Promise.resolve([]),
       };
     });
